backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
# Loose pin: strict == can conflict with extras resolution; LangChain stack may pull httpx newer than 0.25.x
uvicorn[standard]>=0.24.0,<0.35.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# LangGraph and LangChain
# langgraph.types (interrupt, Command) exists in langgraph>=0.2.
# Pin a known-compatible LangChain 0.2 set to avoid pip resolver conflicts.
langgraph==0.2.48
langchain-core==0.2.43
langchain==0.2.17
langchain-openai==0.1.25
langchain-anthropic==0.1.23
langchain-community==0.2.17
langchain-google-genai==1.0.10
langsmith>=0.1.112,<0.2.0

# Gmail API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx>=0.25.0,<0.29.0
PyJWT==2.8.0
slowapi==0.1.9
cryptography>=42.0.0,<45.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
